server/requirements/requirements.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
retrying==1.3.3
six==1.11.0
SQLAlchemy==1.2.9
Werkzeug==0.14.1
deprecation==2.0.6
xlsxwriter==1.1.1
MySQL-connector==2.1.6
MySQL-python==1.2.5
pyjwt==1.7.1
cryptography==2.4.2
